Break out some controllers and move mail to subpackage
Controllers moved into sub-packages:
* Initial setup
* Admin routes
* Auth login/out routes
* Contact
