+++ root/account.gohtml
@@ -2,6 +2,7 @@
 {{ define "content" }}
 <h1>Account Settings</h1>
 
+{{ $form := .Data.Form }}
 <form action="/account" method="POST">
     <input type="hidden" name="_csrf" value="{{ .CSRF }}">
 
@@ -14,7 +15,7 @@
             class="form-control"
             name="username"
             id="username"
-            value="{{ .Form.Username }}"
+            value="{{ $form.Username }}"
             placeholder="soandso">
     </div>
 
@@ -24,8 +25,8 @@
             class="form-control"
             name="name"
             id="name"
-            value="{{ .Form.Name }}"
-            placeholder="{{ or .Form.Username "Anonymous" }}">
+            value="{{ $form.Name }}"
+            placeholder="{{ or $form.Username "Anonymous" }}">
     </div>
 
     <div class="form-group">
@@ -34,7 +35,7 @@
             class="form-control"
             name="email"
             id="email"
-            value="{{ .Form.Email }}"
+            value="{{ $form.Email }}"
             placeholder="name@domain.com">
     </div>
