+++ root/contact.gohtml
@@ -7,8 +7,6 @@
     administrator.
 </p>
 
-data={{ .Data }}
-
 {{ $form := .Data.Form }}
 <form method="POST" action="/contact">
     <input type="hidden" name="_csrf" value="{{ .CSRF }}">
